pyproject.toml
[project]
name = "neopixel-pi5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "viam-sdk>=0.55.1",
]
